Move files to separate folders
