+++ c_cxx/accuracy_tool/CMakeLists.txt
@@ -60,19 +60,22 @@ else()
  string(APPEND CMAKE_C_FLAGS " -Wall -Wextra")
 endif()
 
-add_executable(accuracy_test main.cc
-                             cmd_args.h
-                             cmd_args.cc
-                             basic_utils.h
-                             basic_utils.cc
-                             model_io_utils.h
-                             model_io_utils.cc
-                             data_loader.h
-                             data_loader.cc
-                             acc_task.h
-                             acc_task.cc
-                             task_thread_pool.h
-                             task_thread_pool.cc)
+add_executable(accuracy_test src/main.cc
+                             src/cmd_args.h
+                             src/cmd_args.cc
+                             src/ep_cmd_args/qnn_cmd_args.h
+                             src/ep_cmd_args/qnn_cmd_args.cc
+                             src/basic_utils.h
+                             src/basic_utils.cc
+                             src/model_io_utils.h
+                             src/model_io_utils.cc
+                             src/data_loader.h
+                             src/data_loader.cc
+                             src/acc_task.h
+                             src/acc_task.cc
+                             src/task_thread_pool.h
+                             src/task_thread_pool.cc)
+target_include_directories(accuracy_test PUBLIC "${PROJECT_SOURCE_DIR}/src")
 
 find_package(Threads REQUIRED)
 target_link_libraries(accuracy_test onnxruntime Threads::Threads)
