+++ c_cxx/accuracy_tool/src/basic_utils.cc
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 #include "basic_utils.h"
+
 #include <fstream>
 #include <string>
