+++ c_cxx/accuracy_tool/src/data_loader.cc
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 #include "data_loader.h"
+
 #include <algorithm>
 #include <cassert>
 #include <filesystem>
@@ -10,10 +11,8 @@
 
 namespace acctest {
 
-bool LoadIODataFromDisk(const std::vector<std::filesystem::path>& dataset_paths,
-                        const std::vector<IOInfo>& io_infos,
-                        const char* data_file_prefix,
-                        std::vector<std::unique_ptr<char[]>>& dataset_data) {
+bool LoadIODataFromDisk(const std::vector<std::filesystem::path>& dataset_paths, const std::vector<IOInfo>& io_infos,
+                        const char* data_file_prefix, std::vector<std::unique_ptr<char[]>>& dataset_data) {
   size_t total_data_size = 0;
   for (const auto& io_info : io_infos) {
     total_data_size += io_info.total_data_size;
@@ -49,8 +48,7 @@ bool LoadIODataFromDisk(const std::vector<std::filesystem::path>& dataset_paths,
       const size_t io_index = static_cast<size_t>(io_index_s32);
       if (io_index >= io_infos.size()) {
         std::cerr << "[ERROR]: The input (or output) file index for file " << data_file_path
-                  << " exceeds the number of inputs (or outputs) in the model ("
-                  << io_infos.size() << ")" << std::endl;
+                  << " exceeds the number of inputs (or outputs) in the model (" << io_infos.size() << ")" << std::endl;
         return false;
       }
 
@@ -70,9 +68,9 @@ bool LoadIODataFromDisk(const std::vector<std::filesystem::path>& dataset_paths,
     }
 
     if (num_files_loaded != io_infos.size()) {
-      std::cerr << "[ERROR]: " << dataset_path << " does not have the expected number of "
-                << data_file_prefix << "<i>.raw files. Loaded " << num_files_loaded << "files, but expected "
-                << io_infos.size() << "files." << std::endl;
+      std::cerr << "[ERROR]: " << dataset_path << " does not have the expected number of " << data_file_prefix
+                << "<i>.raw files. Loaded " << num_files_loaded << " files, but expected " << io_infos.size()
+                << "files." << std::endl;
       return false;
     }
   }
