+++ c_cxx/accuracy_tool/src/ep_cmd_args/qnn_cmd_args.h
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+#pragma once
+#include "cmd_args.h"
+
+bool ParseQnnEpArgs(AppArgs& app_args, CmdArgParser& cmd_args);
