+++ c_cxx/accuracy_tool/src/task_thread_pool.h
@@ -7,8 +7,8 @@
 #include <thread>
 #include <vector>
 
-#include "basic_utils.h"
 #include "acc_task.h"
+#include "basic_utils.h"
 
 class TaskThreadPool {
  public:
